Move links validation code to its own file (#326)
