+++ code/go/internal/validator/semantic/validate_version_integrity_test.go
